+++ src/app/shared/enums/obtaining-method.enum.ts
@@ -1,0 +1,9 @@
+export enum ObtainingMethods {
+  UNKNOWN = 1,
+  CAPTURE,
+  EVOLVING,
+  BREEDING,
+  TRADING,
+  EVENT,
+  GIFT
+}