+++ src/app/shared/interfaces/shiny-list.interface.ts
@@ -1,0 +1,5 @@
+import { ShinyDexEntry } from '@shared/interfaces/shiny-dex-entry.interface';
+
+export interface ShinyList {
+  [key: string]: ShinyDexEntry;
+}