+++ src/app/tools/pipes/shiny-dex-entry-to-pokemon.pipe.ts
@@ -1,16 +1,21 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { ShinyDexEntry } from '../models/shiny-dex-entry.interface';
 import { Pokemon } from '@shared/interfaces/pokemon';
+import { ShinyDexEntry } from '@shared/interfaces/shiny-dex-entry.interface';
 
 @Pipe({
   name: 'shinyDexEntryToPokemon'
 })
 export class ShinyDexEntryToPokemonPipe implements PipeTransform {
 
-  transform(value: ShinyDexEntry): Pokemon {
+  transform(value: ShinyDexEntry, options?: Record<string, unknown>): Pokemon {
     const pokemon = value.pokemon as unknown as Pokemon;
     pokemon.isShiny = true;
     pokemon.isOwned = !!value.amountShinies;
+    if (options) {
+      if (typeof options.item === 'string') {
+        pokemon.item = options.item;
+      }
+    }
     return pokemon;
   }