+++ src/app/tools/tools-routing.module.ts
@@ -3,6 +3,10 @@ import { RouterModule, Routes } from '@angular/router';
 import { ToolsComponent } from './tools.component';
 import { BreedingComponent } from './breeding/breeding.component';
 import { ShinyDexComponent } from './components/shiny-dex/shiny-dex.component';
+import { FriendSafariShinyDexComponent } from './components/friend-safari-shiny-dex/friend-safari-shiny-dex.component';
+import {
+  UltraWormholeShinyDexComponent
+} from './components/ultra-wormhole-shiny-dex/ultra-wormhole-shiny-dex.component';
 
 
 const routes: Routes = [
@@ -16,6 +20,12 @@ const routes: Routes = [
       },
       {
         path: 'shiny-dex', component: ShinyDexComponent
+      },
+      {
+        path: 'friend-safari-shiny-dex', component: FriendSafariShinyDexComponent
+      },
+      {
+        path: 'ultra-wormhole-shiny-dex', component: UltraWormholeShinyDexComponent
       }
     ]
   }